design: 전체 모임 레이아웃 재구성(가이드, 버튼 등 위치 조정)
