+++ src/components/page/list/Filter/index.tsx
@@ -2,8 +2,6 @@ import { styled } from 'stitches.config';
 import { Flex } from '@components/util/layout/Flex';
 import { useSearchParams } from '@hooks/queryString/custom';
 import Search from './Search';
-import ArrowRightCircleIcon from '@assets/svg/arrow_right_circle.svg';
-import Toggle from './Modal/Toggle';
 import { CATEGORY_FILTER, GENERATION_FILTER, PART_FILTER, STATUS_FILTER } from '@constants/option';
 import DropDownFilter from './DropDown';
 
@@ -14,21 +12,13 @@ function Filter() {
     <>
       <Flex align="center" justify="between">
         <Flex>
-          <Search />
           <DropDownFilter filter={CATEGORY_FILTER} />
           <DropDownFilter filter={PART_FILTER} />
           <DropDownFilter filter={STATUS_FILTER} />
           <DropDownFilter filter={GENERATION_FILTER} />
         </Flex>
 
-        <SGuideButton
-          target="_blank"
-          href="https://www.notion.so/sopt-makers/eec46a4562ec48f0b0220153bb6ea68e"
-          rel="noreferrer noopener"
-        >
-          모임 신청 가이드
-          <ArrowRightCircleIcon />
-        </SGuideButton>
+        <Search />
       </Flex>
 
       {!!search && <SearchResultMessage>"{search}"에 대한 검색결과입니다.</SearchResultMessage>}
@@ -37,26 +27,6 @@ function Filter() {
 }
 
 export default Filter;
-
-const SGuideButton = styled('a', {
-  height: '$48',
-  flexType: 'verticalCenter',
-  gap: '$8',
-  color: '$gray10',
-  padding: '$18 $20',
-  borderRadius: '14px',
-  fontAg: '18_semibold_100',
-  boxSizing: 'border-box',
-  '@tablet': {
-    padding: '$14 $12 $14 $16',
-    borderRadius: '10px',
-    fontAg: '14_medium_100',
-  },
-
-  path: {
-    stroke: '$gray10',
-  },
-});
 
 const SearchResultMessage = styled('p', {
   fontAg: '24_medium_100',
